+++ src/components/HomeScreen.js
@@ -1,10 +1,18 @@
 import React, { Component } from 'react';
-import { StyleSheet, Text, View, SafeAreaView } from 'react-native';
+import {
+  StyleSheet,
+  Text,
+  View,
+  SafeAreaView,
+  TouchableOpacity,
+  Image,
+} from 'react-native';
 import { Texts } from '../util/constants/Strings';
 import { Colors, NavigationConstants } from '../util/constants/Constants';
 import CustomMainButton from '../util/CustomMainButton';
 import { Data } from './MockData';
-import VoiceTest from '../util/VoiceTest';
+import Voice from 'react-native-voice';
+import { ImageConstants } from '../assets/ImageConstants';
 
 export default class HomeScreen extends Component {
   constructor(props) {
@@ -12,6 +20,13 @@ export default class HomeScreen extends Component {
     this.state = {
       dishSelected: '',
     };
+    Voice.onSpeechStart = this.onSpeechStart;
+    Voice.onSpeechRecognized = this.onSpeechRecognized;
+    Voice.onSpeechEnd = this.onSpeechEnd;
+    Voice.onSpeechError = this.onSpeechError;
+    Voice.onSpeechResults = this.onSpeechResults;
+    Voice.onSpeechPartialResults = this.onSpeechPartialResults;
+    Voice.onSpeechVolumeChanged = this.onSpeechVolumeChanged;
   }
 
   buttonPressed = (title) => {
@@ -22,13 +37,100 @@ export default class HomeScreen extends Component {
       value: title,
     });
   };
+  async componentDidMount() {
+    this.startRecognizing();
+  }
+
+  async componentWillUnmount() {
+    this.destroyRecognizer();
+    Voice.destroy().then(Voice.removeAllListeners);
+  }
+
+  async startRecognizing() {
+    this.setState({
+      result: ' ',
+    });
+
+    try {
+      await Voice.start('en-US');
+    } catch (e) {
+      console.error(e);
+    }
+  }
+
+  async destroyRecognizer() {
+    try {
+      await Voice.destroy();
+    } catch (e) {
+      console.error(e);
+    }
+    this.setState({
+      results: ' ',
+    });
+  }
+
+  onSpeechStart = (e) => {
+    console.log('onSpeechStart', e);
+  };
+
+  onSpeechRecognized = (e) => {
+    console.log('onSpeechRecognized', e);
+  };
+
+  onSpeechEnd = (e) => {
+    console.log('onSpeechEnd', e);
+  };
+
+  onSpeechError = (e) => {
+    console.log('onSpeechError', e);
+    this.setState({
+      error: JSON.stringify(e.error),
+    });
+  };
+
+  onSpeechPartialResults = (e) => {
+    console.log('onSpeechPartialResults: ', e);
+    this.setState({
+      partialResults: e.value,
+    });
+  };
+
+  onSpeechVolumeChanged = (e) => {
+    console.log('onSpeechVolumeChanged: ', e);
+  };
+
+  onSpeechResults = (e) => {
+    console.log('onSpeechResults', e);
+    const latestArray = e.value[e.value.length - 1];
+    const indexOfTrigger = latestArray.lastIndexOf('Hey');
+    const question = latestArray
+      .substring(indexOfTrigger + 3, latestArray.length)
+      .toLowerCase();
+    if (question.includes(Data[0].title.toLowerCase())) {
+      this.buttonPressed(Data[0].title);
+    } else if (question.includes(Data[1].title.toLowerCase())) {
+      this.buttonPressed(Data[1].title);
+    } else if (question.includes(Data[2].title.toLowerCase())) {
+      this.buttonPressed(Data[2].title);
+    } else {
+      console.log('Not recognized');
+    }
+  };
 
   render() {
     return (
       <SafeAreaView style={styles.rootContainer}>
         <View style={styles.headingContainer}>
           <Text style={styles.headingText}>{Texts.clove}</Text>
-          <VoiceTest />
+          <View style={styles.micContainer}>
+            <TouchableOpacity
+              onPress={() => {
+                this.startRecognizing();
+              }}
+            >
+              <Image style={styles.img} source={ImageConstants.micIcon} />
+            </TouchableOpacity>
+          </View>
         </View>
         <View style={styles.container}>
           <View>
@@ -92,5 +194,13 @@ const styles = StyleSheet.create({
     color: Colors.darkPurple,
     marginBottom: 20,
     marginLeft: 10,
+  },
+  micContainer: {
+    alignContent: 'center',
+    alignItems: 'center',
+  },
+  img: {
+    width: 50,
+    height: 50,
   },
 });
